+++ src/components/properties.clj
@@ -4,26 +4,20 @@
 (defn create_entity
   "Gives the entity a component value and puts in the data structure."
   [& {:keys [entity, components]}]
-  (let[ entity (transient entity)
-        entity_id (entity :id)
- 	entity_components (entity :components)
- 	new_components (conj entity_components components)]
-    (assoc! entity :components new_components)
-    (persistent! entity)))
-	 
+  (let [tentity (transient entity)
+    entity-components (:components entity)
+    new-components (conj (or entity-components #{}) components)
+    updated (assoc! tentity :components new-components)]
+    (persistent! updated)))
 
 (defn get_properties
   "Get properties of a given entity."
   [entity]
-  (let[ entity (transient entity)
-        entity_id (entity :id)
-        entity_components (entity :components)]
-    (prn entity_components))
-  )
+  (:components entity))
 
 (defn set_component_configuration
   "Configure an entity's components according to component change parameters"
   [entity component_changes]
-  
-  (prn "ok")
-)
+  (let [existing (:components entity)
+    new-components (conj (or existing #{}) component_changes)]
+    (assoc entity :components new-components)))